[616] PR feedback
