--- app/components/trainees/assessment_details/view.html.erb
@@ -1,14 +1,0 @@
-<%= render_component SummaryCard::View.new(
-  title: "Assessment details",
-  heading_level: 2,
-  rows: [
-    {
-      key: "Assessment outcome",
-      value: assessment_outcome,
-    },
-    {
-      key: "Assessment end date",
-      value: assessment_end_date,
-    },
-  ])
-%>
