+++ app/components/trainees/outcome_details/view.html.erb
@@ -1,0 +1,10 @@
+<%= render_component SummaryCard::View.new(
+  title: "Outcome details",
+  heading_level: 2,
+  rows: [
+    {
+      key: "Date standards met",
+      value: outcome_date,
+    },
+  ])
+%>
